requirements.txt
PyYAML==6.0
Flask==3.0.0
pytest-cov
coverage